Refactored study_utils
