+++ lib/core/routing/routes.dart
@@ -22,13 +22,13 @@ import 'package:kanpractice/ui/pages/study_modes/recognition.dart';
 import 'package:kanpractice/ui/pages/settings/settings.dart';
 import 'package:kanpractice/ui/pages/statistics/statistics.dart';
 import 'package:kanpractice/ui/pages/study_modes/speaking.dart';
+import 'package:kanpractice/ui/pages/study_modes/utils/mode_arguments.dart';
 import 'package:kanpractice/ui/pages/test_history/bloc/test_bloc.dart';
 import 'package:kanpractice/ui/pages/test_history/test_history.dart';
 import 'package:kanpractice/ui/pages/test_result/arguments.dart';
 import 'package:kanpractice/ui/pages/test_result/test_result.dart';
 import 'package:kanpractice/ui/pages/tutorial/tutorial.dart';
 import 'package:kanpractice/ui/pages/study_modes/writing.dart';
-import 'package:kanpractice/core/utils/study_modes/mode_arguments.dart';
 import 'package:kanpractice/ui/pages/word_history/bloc/word_history_bloc.dart';
 import 'package:kanpractice/ui/pages/word_history/word_history.dart';
 import 'package:page_transition/page_transition.dart';
