+++ lib/ui/pages/study_modes/listening.dart
@@ -8,8 +8,8 @@ import 'package:kanpractice/ui/general_utils.dart';
 import 'package:kanpractice/core/utils/tts.dart';
 import 'package:kanpractice/core/types/study_modes.dart';
 import 'package:kanpractice/ui/consts.dart';
-import 'package:kanpractice/core/utils/study_modes/mode_arguments.dart';
-import 'package:kanpractice/core/utils/study_modes/study_mode_update_handler.dart';
+import 'package:kanpractice/ui/pages/study_modes/utils/mode_arguments.dart';
+import 'package:kanpractice/ui/pages/study_modes/utils/study_mode_update_handler.dart';
 import 'package:kanpractice/ui/widgets/kp_learning_header_animation.dart';
 import 'package:kanpractice/ui/widgets/kp_learning_header_container.dart';
 import 'package:kanpractice/ui/widgets/kp_list_percentage_indicator.dart';
