+++ lib/ui/pages/study_modes/utils/study_mode_update_handler.dart
@@ -6,8 +6,8 @@ import 'package:kanpractice/core/database/queries/list_queries.dart';
 import 'package:kanpractice/core/preferences/store_manager.dart';
 import 'package:kanpractice/core/routing/pages.dart';
 import 'package:kanpractice/core/types/study_modes.dart';
+import 'package:kanpractice/ui/pages/study_modes/utils/mode_arguments.dart';
 import 'package:kanpractice/ui/pages/test_result/arguments.dart';
-import 'package:kanpractice/core/utils/study_modes/mode_arguments.dart';
 import 'package:kanpractice/ui/widgets/kp_alert_dialog.dart';
 import 'package:easy_localization/easy_localization.dart';
