+++ lib/ui/widgets/blitz/kp_number_test_bottom_sheet.dart
@@ -6,7 +6,7 @@ import 'package:kanpractice/core/routing/pages.dart';
 import 'package:kanpractice/core/types/number_ranges.dart';
 import 'package:kanpractice/core/types/test_modes.dart';
 import 'package:kanpractice/ui/general_utils.dart';
-import 'package:kanpractice/core/utils/study_modes/mode_arguments.dart';
+import 'package:kanpractice/ui/pages/study_modes/utils/mode_arguments.dart';
 import 'package:kanpractice/core/types/study_modes.dart';
 import 'package:kanpractice/ui/widgets/kp_button.dart';
 import 'package:kanpractice/ui/widgets/kp_drag_container.dart';
